Begin work on rewriting the parser
Completes the first step of work in porting the Nom parser to a hand-written
recursive descent parser.  This initial patch implements a fast lexer that
attempts to avoid allocations by returning tokens that contain references to
the original source data and starting/ending character positions of the lexeme.

Additionally, the AST has been completely refactored and now has support for tracking
locational information for each node in the AST from top-level nodes down to operators.